Revert "Allow user to undo icon deletions"
This reverts commit d1471157c4230a3fbcfdf9401c22e2d4c104b782.
