--- lawnchair/res/layout/snackbar.xml
@@ -1,52 +1,0 @@
-<?xml version="1.0" encoding="utf-8"?>
-<!--
-  ~     This file is part of Lawnchair Launcher.
-  ~
-  ~     Lawnchair Launcher is free software: you can redistribute it and/or modify
-  ~     it under the terms of the GNU General Public License as published by
-  ~     the Free Software Foundation, either version 3 of the License, or
-  ~     (at your option) any later version.
-  ~
-  ~     Lawnchair Launcher is distributed in the hope that it will be useful,
-  ~     but WITHOUT ANY WARRANTY; without even the implied warranty of
-  ~     MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the
-  ~     GNU General Public License for more details.
-  ~
-  ~     You should have received a copy of the GNU General Public License
-  ~     along with Lawnchair Launcher.  If not, see <https://www.gnu.org/licenses/>.
-  -->
-
-<merge xmlns:android="http://schemas.android.com/apk/res/android"
-    xmlns:app="http://schemas.android.com/apk/res-auto"
-    android:layout_width="fill_parent"
-    android:layout_height="fill_parent">
-    <TextView
-        android:id="@id/label"
-        android:layout_width="0.0dip"
-        android:layout_height="@dimen/snackbar_content_height"
-        android:layout_weight="1"
-        android:paddingLeft="8.0dip"
-        android:paddingRight="8.0dip"
-        android:ellipsize="end"
-        android:gravity="center_vertical"
-        android:lines="1"
-        android:textColor="?android:textColorPrimary"
-        android:textSize="@dimen/snackbar_max_text_size"
-        android:theme="@style/TextTitle"
-        app:customFontType="text"/>
-    <TextView
-        android:id="@+id/action"
-        android:textStyle="bold"
-        android:layout_width="wrap_content"
-        android:layout_height="@dimen/snackbar_content_height"
-        android:layout_weight="0"
-        android:paddingLeft="8.0dip"
-        android:paddingRight="8.0dip"
-        android:background="?android:selectableItemBackground"
-        android:capitalize="sentences"
-        android:gravity="center"
-        android:textColor="?android:colorAccent"
-        android:textSize="@dimen/snackbar_max_text_size"
-        android:theme="@style/TextTitle"
-        app:customFontType="button"/>
-</merge>
