+++ src/com/android/launcher3/DeleteDropTarget.java
@@ -21,12 +21,10 @@ import android.text.TextUtils;
 import android.util.AttributeSet;
 import android.view.View;
 
-import ch.deletescape.lawnchair.views.Snackbar;
 import com.android.launcher3.accessibility.LauncherAccessibilityDelegate;
 import com.android.launcher3.dragndrop.DragOptions;
 import com.android.launcher3.folder.Folder;
 import com.android.launcher3.logging.LoggerUtils;
-import com.android.launcher3.model.ModelWriter;
 import com.android.launcher3.userevent.nano.LauncherLogProto.ControlType;
 import com.android.launcher3.userevent.nano.LauncherLogProto.Target;
 
@@ -102,14 +100,7 @@ public class DeleteDropTarget extends ButtonDropTarget {
     public void completeDrop(DragObject d) {
         ItemInfo item = d.dragInfo;
         if ((d.dragSource instanceof Workspace) || (d.dragSource instanceof Folder)) {
-            mLauncher.getModelWriter().prepareToUndo();
-
             onAccessibilityDrop(null, item);
-
-            int currentPage = this.mLauncher.getWorkspace().getCurrentPage();
-            Snackbar.show(mLauncher, R.string.item_removed, R.string.undo,
-                    () -> mLauncher.getModelWriter().commitDelete(),
-                    () -> mLauncher.getModelWriter().undoDelete(currentPage));
         }
     }
