+++ src/com/android/launcher3/Launcher.java
@@ -1665,7 +1665,18 @@ public class Launcher extends BaseDraggingActivity implements LauncherExterns,
      * Deletes the widget info and the widget id.
      */
     private void deleteWidgetInfo(final LauncherAppWidgetInfo widgetInfo) {
-        getModelWriter().deleteWidgetInfo(widgetInfo, getAppWidgetHost());
+        final LauncherAppWidgetHost appWidgetHost = getAppWidgetHost();
+        if (appWidgetHost != null && !widgetInfo.isCustomWidget() && widgetInfo.isWidgetIdAllocated()) {
+            // Deleting an app widget ID is a void call but writes to disk before returning
+            // to the caller...
+            new AsyncTask<Void, Void, Void>() {
+                public Void doInBackground(Void ... args) {
+                    appWidgetHost.deleteAppWidgetId(widgetInfo.appWidgetId);
+                    return null;
+                }
+            }.executeOnExecutor(Utilities.THREAD_POOL_EXECUTOR);
+        }
+        getModelWriter().deleteItemFromDatabase(widgetInfo);
     }
 
     @Override
