+++ src/com/android/launcher3/LauncherModel.java
@@ -439,11 +439,7 @@ public class LauncherModel extends BroadcastReceiver
      * not be called as DB updates are automatically followed by UI update
      */
     public void forceReload() {
-        forceReload(-1);
-    }
-
-    public void forceReload(int page) {
-        synchronized (this.mLock) {
+        synchronized (mLock) {
             // Stop any existing loaders first, so they don't set mModelLoaded to true later
             stopLoader();
             mModelLoaded = false;
@@ -451,12 +447,11 @@ public class LauncherModel extends BroadcastReceiver
 
         // Start the loader if launcher is already running, otherwise the loader will run,
         // the next time launcher starts
-        Callbacks callback = getCallback();
-        if (callback != null) {
-            startLoader(page < 0 ? callback.getCurrentWorkspaceScreen() : page);
+        Callbacks callbacks = getCallback();
+        if (callbacks != null) {
+            startLoader(callbacks.getCurrentWorkspaceScreen());
         }
     }
-
 
     public boolean isCurrentCallbacks(Callbacks callbacks) {
         return (mCallbacks != null && mCallbacks.get() == callbacks);
